Enabling more than one connection for H2 file based DB
In development H2 will be used as a file based datbase.
By default file based databases allows only one connection.
If you want to open another connection, the old one
must be closed before opening a new one.

This safety feature is necessary for because in file based
databases there are no connection pool.

In development this safety features are not that important
and it slows down the development process to shutdown
the webserver every time I want to access the database
by an SQL explorer like DBeaver.

So keep in mind, that this feature is only for development
and you need to set the same parameter in you SQL Explorer
too.