REST API for GET /containers now sends all currentMeasures too
In the default installation of jhipster no lists will be sent.
To avoid this I loop through all measures (lazy loading) and set
the container in all measures to null to avoid an infinite loop.

This is not a clean solution but it works. If anyone knows a
better solution. I appreciate any help.